Refactor Form component and use new folder structure
